added support for armor to contain modifiers

+++ src/main/java/com/wdiscute/starcatcher/registry/custom/catchmodifiers/AbstractCatchModifier.java
@@ -31,7 +31,7 @@ public abstract class AbstractCatchModifier
     }
 
     //server only
-    public int adjustChanceToFishEachTick(int chanceToFishEachTick)
+    public float adjustChanceToFishEachTick(float chanceToFishEachTick)
     {
         return chanceToFishEachTick;
     }
